+++ pickLayer/plugin.py
@@ -133,7 +133,7 @@ class Plugin:
     def unload(self) -> None:
         """Removes the plugin menu item and icon from QGIS GUI."""
         for action in self.actions:
-            iface.removePluginMenu(plugin_name(), action)
+            iface.removePluginMenu(self.menu, action)
             iface.removeToolBarIcon(action)
             iface.unregisterMainWindowAction(action)